requirements.txt
python-telegram-bot>=21,<22
ollama>=0.3,<1
chromadb>=0.5,<1
sqlalchemy>=2,<3
aiosqlite>=0.20
apscheduler>=3.10,<4
dateparser>=1.2
httpx>=0.27,<1
python-dotenv>=1
structlog>=24
